seperated register and log in pages and began setting up the user profile html -Dan
