+++ public/index.html
@@ -5,29 +5,23 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="./assets/css/style.css">
-    <title>Login Page</title>
+    <title>Register Page</title>
 </head>
 <body>
     <div class="login"> 
-        <form class="form" id="login-form">
-            <div id="login-text">Login</div>
-            <input class="input" type="text" id="username-input" placeholder="username">
-            <input class="input" type="text" id="password-input" placeholder="password">
-            <button>Submit</button>
-        </form>  
         <form class="form" id="register-form">
-            <div id="register-text">Register</div>
-            <input class="input" type="text" id="first-name-input" placeholder="first name">
-            <input class="input" type="text" id="last-name-input" placeholder="last name">
-            <input class="input" type="text" id="email-input" placeholder="email">
-            <input class="input" type="text" id="username-input" placeholder="username">
-            <input class="input" type="text" id="password-input" placeholder="password">
-            <button>Submit</button>
+            <div class="big-text" id="register-text">Register</div>
+            <div class="center-input">            
+                <input class="input" type="text" id="first-name-input" placeholder="first name">
+                <input class="input" type="text" id="last-name-input" placeholder="last name">
+                <input class="input" type="text" id="email-input" placeholder="email">
+                <input class="input" type="text" id="username-input" placeholder="username">
+                <input class="input" type="text" id="password-input" placeholder="password"></div>
+
+            <button class="btn" id="register-btn">Register</button>
+            <div class="login-link">Already have an account?<a href="./login.html"> Log in.</a></div>
         </form> 
     </div>
-    <div class="register">
-        <form class="register-form"></form>
-    </div>
-    <a class="footing" href="http://www.freepik.com">Designed by gstudioimagen / Freepik</a>
+    <a class="footing" href="http://www.freepik.com">Background Designed by gstudioimagen / Freepik</a>
 </body>
 </html>
