+++ public/login.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="./assets/css/style.css">
+    <title>Login Page</title>
+</head>
+<body>
+    <div class="login"> 
+        <form class="form" id="login-form">
+            <div class="big-text" id="login-text">Login</div>
+            <div class="center-input">
+                <input class="input" type="text" id="username-input" placeholder="username">
+                <input class="input" type="text" id="password-input" placeholder="password">
+            </div>
+            <button class="btn" id="login-btn">Login</button>
+            
+        </form>  
+    </div>
+    <a class="footing" href="http://www.freepik.com">Designed by gstudioimagen / Freepik</a>
+</body>
+</html>
