+++ public/user-profile.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="./assets/css/user.css">
+    <title>User Profile</title>
+</head>
+<body>
+    <div class="navbar">
+        <button class="home-btn"><i></i></button>>
+    </div>
+    <div class="flexbox-container1">
+        <div class="flexbox-profile-pic"></div>
+        <div class="flexbox-info"></div>
+    </div>
+    <div class="flexbox-container2">
+        <div class="flexbox-game-list">
+            <div class="game"></div>
+            <div class="game"></div>
+            <div class="game"></div>
+            <div class="game"></div>
+            <div class="game"></div>
+            <div class="game"></div>
+        </div>
+    </div>
+</body>
+</html>
